LayoutTests/http/tests/xmlhttprequest/workers/shared-worker-access-control-basic-get-fail-non-simple-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-basic-get-fail-non-simple.cgi. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
GET should not trigger a preflight request from a shared worker unless it has non-simple headers.

PASS: Cross-domain access allowed for simple get.
PASS: Exception thrown. Cross-domain access was denied in 'send'. [NETWORK_ERR: XMLHttpRequest Exception 101].
DONE
